+++ bigbluebutton-html5/imports/ui/components/actions-bar/component.jsx
@@ -5,7 +5,7 @@ import DesktopShare from './desktop-share/component';
 import ActionsDropdown from './actions-dropdown/component';
 import QuickPollDropdown from './quick-poll-dropdown/component';
 import AudioControlsContainer from '../audio/audio-controls/container';
-import JoinVideoOptionsContainer from '../multiple-video-provider/video-button/container';
+import JoinVideoOptionsContainer from '../video-provider/video-button/container';
 import CaptionsButtonContainer from '/imports/ui/components/actions-bar/captions/container';
 import PresentationOptionsContainer from './presentation-options/component';
