--- bigbluebutton-html5/imports/ui/components/multiple-video-provider/video-list/video-list-item/container.jsx
@@ -1,18 +1,0 @@
-import React from 'react';
-import { withTracker } from 'meteor/react-meteor-data';
-import VoiceUsers from '/imports/api/voice-users/';
-import VideoListItem from './component';
-
-const VideoListItemContainer = props => (
-  <VideoListItem {...props} />
-);
-
-export default withTracker((props) => {
-  const {
-    userId,
-  } = props;
-
-  return {
-    voiceUser: VoiceUsers.findOne({ intId: userId }),
-  };
-})(VideoListItemContainer);
