--- bigbluebutton-html5/imports/ui/components/video-provider/styles.scss
@@ -1,18 +1,0 @@
-// .webcam {
-//   position: relative;
-//
-//   &:before {
-//	 display: block;
-//	 content: "";
-//	 width: 100%;
-//	 padding-top: (var(--height) / var(--width)) * 100%;
-//   }
-//
-//   > .content {
-//	 position: absolute;
-//	 top: 0;
-//	 left: 0;
-//	 right: 0;
-//	 bottom: 0;
-//   }
-// }
