+++ bigbluebutton-html5/imports/ui/components/video-provider/video-button/container.jsx
@@ -3,43 +3,31 @@ import { withTracker } from 'meteor/react-meteor-data';
 import { injectIntl } from 'react-intl';
 import { withModalMounter } from '/imports/ui/components/modal/service';
 import VideoPreviewContainer from '/imports/ui/components/video-preview/container';
-import { notify } from '/imports/ui/services/notification';
 import JoinVideoButton from './component';
-import VideoButtonService from './service';
-
-import {
-  validIOSVersion,
-} from '/imports/ui/components/app/service';
+import VideoService from '../service';
 
 const JoinVideoOptionsContainer = (props) => {
   const {
-    isSharingVideo,
+    hasVideoStream,
     isDisabled,
     handleJoinVideo,
     handleCloseVideo,
-    baseName,
     intl,
     mountModal,
-    isMobileNative,
     ...restProps
   } = props;
 
   const mountPreview = () => { mountModal(<VideoPreviewContainer fromInterface />); };
 
-  return !isMobileNative && (
-  <JoinVideoButton {...{
-    handleJoinVideo: mountPreview, handleCloseVideo, isSharingVideo, isDisabled, ...restProps,
-  }}
-  />
+  return (
+    <JoinVideoButton {...{
+      handleJoinVideo: mountPreview, handleCloseVideo, hasVideoStream, isDisabled, ...restProps,
+    }}
+    />
   );
 };
 
 export default withModalMounter(injectIntl(withTracker(() => ({
-  baseName: VideoButtonService.baseName,
-  isSharingVideo: VideoButtonService.isSharingVideo(),
-  isDisabled: VideoButtonService.isDisabled() || !Meteor.status().connected,
-  videoShareAllowed: VideoButtonService.videoShareAllowed(),
-  isMobileNative: navigator.userAgent.toLowerCase().includes('bbbnative'),
-  notify,
-  validIOSVersion,
+  hasVideoStream: VideoService.hasVideoStream(),
+  isDisabled: VideoService.isDisabled() || !Meteor.status().connected,
 }))(JoinVideoOptionsContainer)));
